+++ TransactionsParser/src/main/java/com/catdav/transactions/parser/main/TransactionListParser.java
@@ -1,9 +1,10 @@
-package com.catdav.transactions.parser;
+package com.catdav.transactions.parser.main;
 
 import com.catdav.transactions.model.TransactionsContainer;
 
 import java.util.List;
 
 public interface TransactionListParser {
+
     TransactionsContainer parse(List<String> transactionsToBeParsed);
 }
